% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmax_min.R
\name{getmax_min}
\alias{getmax_min}
\title{Función para obtener máximos y minimos de un vector numerico.}
\usage{
getmax_min(serievector)
}
\arguments{
\item{serievector}{, es un vector numerico}
}
\value{
Una lista con los valores máximo y mínimo de los números ingresados.
}
\description{
La función encontrara el valor numerico mayor y el menor
}
\examples{
serie_numeros <- c(1,2,2,3,5,5,6,7,7,7,7)
resultado <- getmax_min(serievector = serie_numeros)
resultado

}
\author{
Equipo
}
