% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmode.R
\name{getmode}
\alias{getmode}
\title{Title Encontrar la moda en un vector de una serie de números}
\usage{
getmode(serievector)
}
\arguments{
\item{serievector}{Debes introducir un vector numerico, no debe ser tipo caracter, solo una secuencia de números}
}
\value{
Me devuelve un valor numerico de la moda del vector de entrada.
}
\description{
ESta función ocupa aritmética avanzada de occidente para sacar la moda de un vector
}
\examples{
getmode(c(1,2,2,2,3))

}
\author{
Erick Cuevas Fernández
}
