% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_heatmap.R
\name{subset_heatmap}
\alias{subset_heatmap}
\title{Crear un subgrupo de la matriz y genera un heatmap}
\usage{
subset_heatmap(x, mediciones = NULL, grupos = NULL)
}
\arguments{
\item{x}{matriz}

\item{mediciones}{character. Lineas de la matriz a graficar.}

\item{grupos}{character. Columnas de la matriz a graficar.}
}
\value{
ComplexHeatmap
}
\description{
Crear un subgrupo de la matriz y genera un heatmap
}
\examples{
mi_ruta <- system.file("extdata", "mi_matriz.csv", package = "runRNAseq")
mi_matriz <- as.matrix(read.csv(mi_ruta, row.names = 1))

subset_heatmap(mi_matriz,
mediciones = c("medicion_a", "medicion_b", "medicion_c"),
grupos = c("grupo_d","grupo_e","grupo_g"))
}
