% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValorMinMax.R
\name{valMinMax}
\alias{valMinMax}
\title{Valor minimo y maximo}
\usage{
valMinMax(vector)
}
\arguments{
\item{vector}{Es un vector de tipo numérico}
}
\value{
Devuelve el valor mas pequeño y más grande del vector
}
\description{
Es una funcion que te imprime el valor minimo y maximo de un vector
}
\examples{
serie_nums <- c(1,1,1,2,3,3,3,4,5,5,5,6,6,7,7)
valMinMax(serie_nums)
}
\author{
Rafael Cervantes

Juan Carlos

Abraham Cruz

Karel Vazquez
}
