% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmode.R
\name{getmode}
\alias{getmode}
\title{Title Encontrar la moda de un vector}
\usage{
getmode(serievector)
}
\arguments{
\item{serievector}{Debes insertar un vector del tipo numerico, no debe ser tipo caracter.}
}
\value{
DEVUELVE  VALORES NUMERICOS
}
\description{
Esta funcion busca obtener la moda de un vector de tipo numerico...
}
\examples{
serie_nums <- c(1,1,1,2,3,3,3,4,5,5,5,6,6,7,7)
getmode(serie_nums)
}
\author{
Karel Vazquez
}
